refactor(schedules): apply procedural-review sweep findings
Six-lens consistency review of the full diff surfaced the remaining
scope-fuzz; all mechanical, no behavior changes except loud-renames:

- agents/schedules now exports SchedulerHandlers, SchedulerPayload, and
  SchedulerEventType (they appear in Scheduler's public signatures) and
  carries a scoped @experimental module banner; SchedulerCallbacks,
  SchedulerHandlers, SchedulerPayload tagged @experimental; Agent's
  lifecycle/scheduler properties tagged @experimental.
- The permanent Agent-only host hooks on Scheduler are loud-named
  __DO_NOT_USE_WILL_BREAK__cleanupRoutePrefix/handleAlarmMemoryLimit,
  matching the repo convention (they are permanent internals, unlike the
  WILL_REMOVE sync-read shims).
- unstable_getSchedulePrompt/unstable_scheduleSchema moved off the new
  agents/schedules/parser entry back to the deprecated agents/schedule
  compat entry, so the new surface ships no born-deprecated aliases.
- Scheduler messages stop recommending Agent-only method names: the
  stale-one-shot warning is host-neutral and the sync-read errors give
  both Agent and standalone guidance; constructor JSDoc no longer
  promises a host-method fallback.
- Deleted the phantom ScheduleStorageRow.retry field, two stale
  @template tags, and a stale resolveRetryConfig comment.
- Tests: env imported from cloudflare:workers in capability-harness;
  console-capture.ts moved beside the fixtures that import it
  (tests/shared stays pool-only); stale lifecycle.test.ts pointer fixed;
  the legacy-parser tests-d check now actually asserts.
- Docs: exclusive-contribution paragraph moved into Shared alarm
  ownership; example imports getCurrentAgent from agents/lifecycle;
  next README marks mcp-client Available; changeset names the
  experimental Agent properties.
