+++ packages/agents/src/schedules/index.ts
@@ -1,8 +1,18 @@
+/**
+ * Persistent scheduling for Lifecycle Objects.
+ *
+ * @experimental The Scheduler surface (`Scheduler`, `SchedulerOptions`, and
+ * the callback-map types) may change before stabilizing. `Schedule`,
+ * `ScheduleCriteria`, and `ScheduleOptions` are shared with Agent's stable
+ * scheduling methods.
+ */
 export { Scheduler } from "./scheduler";
-export type { SchedulerOptions } from "./options";
+export type { SchedulerEventType, SchedulerOptions } from "./options";
 export type {
   Schedule,
   ScheduleCriteria,
   ScheduleOptions,
-  SchedulerCallbacks
+  SchedulerCallbacks,
+  SchedulerHandlers,
+  SchedulerPayload
 } from "./types";
