+++ packages/agents/src/tests/capabilities/console-capture.ts
